Project/phase3/requirements.txt
# Phase 3: Buy vs. Rent RL - Python Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0  # Often required with torch

# Reinforcement Learning
gymnasium>=0.29.0

# Jupyter (optional, for running the demo notebook)
jupyter>=1.0.0
ipython>=7.0.0
ipykernel>=6.0.0

# Optional: For better plots
plotly>=5.0.0

# Optional: For progress bars
tqdm>=4.60.0

# Optional: If you want to use Stable-Baselines3 instead
# stable-baselines3>=2.0.0
